tracker_quaternion
You can find other three-axis algorithms by searching for Euler angles on the website. Because Euler angles have a universal lock problem, they are replaced by quaternions. The quaternion has a reference axis, and the remaining axes are based on the reference axis, so it can be consistent with reality.